install, reboot, disable supported

+++ common/service.sh
@@ -7,7 +7,17 @@
 MODDIR=${0%/*}
 
 # This script will be executed in late_start service mode
-cp /system/app/NfcNci/NfcNci_bak.apk /system/app/NfcNci/NfcNci.apk
-sleep 30
-cp /system/app/NfcNci/NfcNci_align.apk /system/app/NfcNci/NfcNci.apk
-killall com.android.nfc
+
+# restore original apk
+cp "$MODDIR/NfcNci_bak.apk" /system/app/NfcNci/NfcNci.apk
+
+if [ ! -f "$MODDIR/disable" ]; then
+  # wait for nfc service to start
+  sleep 30
+
+  # replace original apk
+  cp "$MODDIR/NfcNci_align.apk" /system/app/NfcNci/NfcNci.apk
+
+  # restart nfc service
+  killall com.android.nfc
+fi
